Re-orgainized and Completed PigLatin GUI
