+++ 02week/pigLatin/index_pigLatin.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link rel="stylesheet" type="text/css" href="./style_pigLatin.css">
+    <title>Welcome to Pig Latin Translater</title>
+</head>
+
+<body>
+    <div id="title">
+        <h1>Pig Latin Translator</h1>
+    </div>
+    <div class="container">
+        <div>
+            <label for="english">Enter English Word:</label>
+            <textarea id="pigLatin-word" class="english" type="text"></textarea>
+        </div>
+        <button id="pigLatin-translate">
+            Translate
+        </button>
+        <div>
+            <label for="pigLatin">Pig Latin Translation:</label>
+            <textarea id="result" class="pigLatin"></textarea>
+        </div>
+    </div>
+
+    <script src="./pigLatin.js"></script>
+</body>
+
+</html>
